move function to the only place it is being used.
