+++ containers/ecr-viewer/src/app/components/BaseFilter.tsx
@@ -7,10 +7,10 @@ import React, {
   useRef,
 } from "react";
 
-import { Button, Label } from "@trussworks/react-uswds";
+import { Button } from "@trussworks/react-uswds";
 import classNames from "classnames";
 
-import { toKebabCase, toSentenceCase } from "@/app/utils/format-utils";
+import { toSentenceCase } from "@/app/utils/format-utils";
 
 import {
   FILTER_CLOSED,
@@ -279,55 +279,6 @@ export const RadioDateOptions = ({
         />
       ))}
     </>
-  );
-};
-
-/**
- *  A custom date input component for selecting a date.
- * @param props - The properties for the CustomDateInput component.
- * @param props.label - The label of the custom date input component.
- * @param props.onDateChange - The function that is called when the date changes.
- * @param props.defaultValue - The default value of the date input.
- * @param props.isRequired - Boolean indicating whether or not the date is required.
- * @param props.minValue - The minimum value of the date input.
- * @returns A JSX element containing a date input field and corresponding label.
- */
-export const CustomDateInput = ({
-  label,
-  onDateChange,
-  defaultValue,
-  isRequired,
-  minValue,
-}: {
-  label: string;
-  onDateChange: (date: string) => void;
-  defaultValue: string;
-  isRequired: boolean;
-  minValue?: string;
-}) => {
-  const today = new Date().toLocaleDateString("en-CA");
-  const id = toKebabCase(label);
-  return (
-    <div>
-      <Label htmlFor={id} className="margin-top-1">
-        {label}
-      </Label>
-      <input
-        id={id}
-        data-testid={id}
-        type="date"
-        className="usa-input width-card margin-top-0 border-base-dark"
-        defaultValue={defaultValue}
-        min={minValue}
-        max={today}
-        required={isRequired}
-        aria-label={label}
-        onChange={(e) => {
-          const date = e.target.value;
-          onDateChange(date);
-        }}
-      />
-    </div>
   );
 };
